Refactor engine into modular components
